Issue #25 - the observer traits now take the function name as a static str
The addition of dependency injection will add more complex borrrowing,
so in prepartaion we should remove unncessary borrows.

First target is the borrowing of the reader, handler and writer functions
and trait objects by the observers who only need to identify them for
tracing purposes. The function name as a string will do for this.